itools: add template match filter
This filter performs template matching in 2x images, a haystack image
and a needle image. It will produce a copy of the haystack image with
a rectangle marking the actual location where it finds the needle in
the haystack.

Note that, if the needle has alpha channel, we use randomness to deal
with needles with alpha channel (see
[here](https://stackoverflow.com/questions/4761940/) for a discussion).
This means that, in that case, the results of successive runs may be
slightly different.

Tested:
```
$ ./python/itools-filter.py --filter match -d -i docs/needle.png docs/lena.compose.jpeg docs/lena.match.jpeg
x0 = 10 y0 = 30
```